requirements.txt
requests
pandas
matplotlib
python-dotenv
upstox-api
